requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database dependencies
motor==3.3.2
pymongo==4.6.0

# Data validation and serialization
pydantic==2.5.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0

# File upload support
python-multipart==0.0.6

# Testing dependencies (optional, for development)
# aiohttp==3.12.13 